# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArtistHub Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feed:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get chronological feed for current user
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [following, local, discovery, custom]
        - name: city
          in: query
          schema:
            type: string
        - name: customFeedId
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/update:
    post:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/unlike:
    post:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/react:
    post:
      operationId: reactToPost
      tags: [posts]
      summary: React to a post with the standard reaction set
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionRequest"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/reaction/remove:
    post:
      operationId: removePostReaction
      tags: [posts]
      summary: Remove current user reaction from a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/repost:
    post:
      operationId: repostPost
      tags: [posts]
      summary: Repost an existing post to the current user's feed
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepostRequest"
      responses:
        "201":
          description: Reposted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/comments:
    get:
      operationId: getPostComments
      tags: [posts]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostComment"
    post:
      operationId: createPostComment
      tags: [posts]
      summary: Comment on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostCommentRequest"
      responses:
        "201":
          description: Updated comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostComment"
  /posts/{postId}/comments/{commentId}:
    delete:
      operationId: deletePostComment
      tags: [posts]
      summary: Delete a post comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostComment"
...
  /custom-feeds:
    get:
      operationId: getCustomFeeds
      tags: [posts]
      summary: Get current user custom feeds
      responses:
        "200":
          description: Custom feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomFeed"
    post:
      operationId: createCustomFeed
      tags: [posts]
      summary: Create custom feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFeedRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFeed"
  /custom-feeds/{feedId}:
    post:
      operationId: updateCustomFeed
      tags: [posts]
      summary: Update custom feed
      parameters:
        - name: feedId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFeedRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFeed"
    delete:
      operationId: deleteCustomFeed
      tags: [posts]
      summary: Delete custom feed
      parameters:
        - name: feedId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        actorSurface:
          type: string
          enum: [personal, artist]
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, friends, private]
        repostOfPostId:
          type: integer
          nullable: true
        likeCount:
          type: integer
        isLiked:
          type: boolean
        reactionCounts:
          $ref: "#/components/schemas/ReactionCounts"
        totalReactionCount:
          type: integer
        currentUserReaction:
          type: string
          nullable: true
          enum: [like, heart, wow, angry]
        repostCount:
          type: integer
        commentCount:
          type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/PostComment"
        originalPost:
          allOf:
            - $ref: "#/components/schemas/Post"
          nullable: true
        author:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
      required: [id, userId, actorSurface, content, visibility, likeCount, isLiked, reactionCounts, totalReactionCount, repostCount, commentCount, comments, createdAt]
    CreatePostRequest:
      type: object
      properties:
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        actorSurface:
          type: string
          enum: [personal, artist]
        repostOfPostId:
          type: integer
          nullable: true
        groupId:
          type: integer
          nullable: true
        visibility:
          type: string
          enum: [public, friends, private]
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMediaInput"
      required: []
    UpdatePostRequest:
      type: object
      properties:
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        actorSurface:
          type: string
          enum: [personal, artist]
        visibility:
          type: string
          enum: [public, friends, private]
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMediaInput"
      required: [content]
    RepostRequest:
      type: object
      properties:
        content:
          type: string
    ReactionRequest:
      type: object
      properties:
        reactionType:
          type: string
          enum: [like, heart, wow, angry]
      required: [reactionType]
...
    FeedResponse:
      allOf:
        - $ref: "#/components/schemas/PostsResponse"
        - type: object
          properties:
            mode:
              type: string
          required: [mode]
...